corrijo error de guardar imagen

+++ models/ImagenForm.php
@@ -24,9 +24,7 @@ class ImagenForm extends Model
             $origen = Yii::getAlias('@uploads/' . $filename);
             $destino = Yii::getAlias('@img/' . $filename);
             $this->imagen->saveAs($origen);
-            // if (file_exists($destino)) :
-            //     unlink($destino);
-            // endif;
+            
             rename($origen, $destino);
             return true;
         } else {
@@ -36,14 +34,20 @@ class ImagenForm extends Model
 
     public function subidaAws($id)
     {
-        $iduser = Yii::$app->user->id;
         $filename = $id . '.' . $this->imagen->extension;
-        $destino = Yii::getAlias('@img/' . $iduser  . '/' . $filename);
+        $destino = Yii::getAlias('@img/' . $filename);
         $aws = Yii::$app->awssdk->getAwsSdk();
         $s3 = $aws->createS3();
-        $amazon = $iduser . '/' . $filename;
-        $s3->putObject([
-                'Bucket'       => 'go00',
+        $amazon = $filename;
+        $bucket = 'go00';
+        $existe = $s3->doesObjectExist('go00', $amazon);
+        if ($existe) :
+            $s3->deleteObject([
+                'Bucket'       => $bucket,
+                'Key'          => $amazon,
+            ]);
+            $s3->putObject([
+                'Bucket'       => $bucket,
                 'Key'          => $amazon,
                 'SourceFile'   => $destino,
                 'ACL'          => 'public-read',
@@ -52,28 +56,37 @@ class ImagenForm extends Model
                     'param1' => 'value 1',
                     'param2' => 'value 2'
                 ]
-        ]);
+            ]);
+        else :
+            $s3->putObject([
+                    'Bucket'       => $bucket,
+                    'Key'          => $amazon,
+                    'SourceFile'   => $destino,
+                    'ACL'          => 'public-read',
+                    'StorageClass' => 'REDUCED_REDUNDANCY',
+                    'Metadata'     => [
+                        'param1' => 'value 1',
+                        'param2' => 'value 2'
+                    ]
+            ]);
+        endif;
 
         return true;
     }
 
     public function borradoLocal($id)
     {
-        unlink($id . '.png');
+        unlink(Yii::getAlias('@img/' . $id . '.png'));
     }
 
     public function descarga($key)
     {
         $aws = Yii::$app->awssdk->getAwsSdk();
         $s3 = $aws->createS3();
-        //get the last object from s3
-        //$object = end($result['Contents']);1
-        // $key = $object['Key'];
         $file = $s3->getObject([
             'Bucket' => 'go00',
             'Key' => $key,
         ]);
         return $file;
-        // save it to disk
     }
 }
